Refactor project structure and rename package to me.codeboris.dungeon
