+++ dungeon-generation-core-spigot/src/main/java/me/codeboris/dungeon/room/Room.java
@@ -1,6 +1,6 @@
-package me.carscupcake.dungeon.room;
+package me.codeboris.dungeon.room;
 
-import me.carscupcake.dungeon.util.Pos2d;
+import me.codeboris.dungeon.util.Pos2d;
 
 import java.util.List;
