+++ dungeon-generation-core-spigot/src/main/java/me/codeboris/dungeon/schematic/SchematicFormat.java
@@ -1,4 +1,4 @@
-package me.carscupcake.dungeon.schematic;
+package me.codeboris.dungeon.schematic;
 
 import org.bukkit.Material;
 import org.bukkit.block.Block;
